Fix errors

+++ 0x03-python-data_structures/9-max_integer.py
@@ -2,7 +2,7 @@
 def max_integer(my_list=[]):
     if len(my_list) == 0:
         return None
-    max_val = int('-inf')
+    max_val = float('-inf')
     for i in my_list:
         max_val = int(i) if int(i) > max_val else max_val
     return max_val
